rename components into views for playground
