+++ src/pages/playground/[play].tsx
@@ -1,6 +1,6 @@
 import dynamic from 'next/dynamic';
 
-const Question = dynamic(() => import('../../playground/components/question'), { ssr: false });
+const Question = dynamic(() => import('../../playground/views/question'), { ssr: false });
 
 export default function Quiz() {
 	return (
