+++ src/pages/playground/index.tsx
@@ -1,4 +1,4 @@
-import Questions from '@/playground/components/allQuestions';
+import Questions from '@/playground/views/allQuestions';
 export default function Quiz() {
 	return (
 		<>
